BE-88, fix : refine jpa mapping dto responses
- Count Condition to Integer ( Long is too big )
- BIg Decimal to Float ( Same reason with Integer )
- WIP : going to fix findByIds
-> according to some articles, finding by ids needs join in queryDsl

+++ src/main/java/io/openur/domain/userchallenge/entity/UserChallengeEntity.java
@@ -40,5 +40,5 @@ public class UserChallengeEntity {
     @Column(nullable = false)
     private Boolean nftCompleted;
 
-    private Long currentCount;
+    private Integer currentCount;
 }
